requirements.txt
requests
beautifulsoup4
colorama
pillow
qrcode
reportlab
flask
pycryptodome
pyelftools
pefile
capstone
python-whois
dnspython
phonenumbers
shodan